component moved to src/lib
- src/lib will contain all the component code
- lib/index.js will be entrypoint to component
- babel will transpile code to `dist/index.js`.
- This will be entrypoint for package
