+++ src/lib/components/Dashboard.js
@@ -3,17 +3,16 @@
  *
  * @author Mayur Borse <mayur@hyphenos.io>
  */
-import React, { useState, useCallback } from "react";
-import defaultConfig from "./constants/defaultConfig";
-import Table from "./Table";
-import PacketDetailsViewer from "./PacketDetailsViewer";
-import "./Dashboard.css";
+import React, { useState, useCallback } from 'react'
+import defaultConfig from '../constants/defaultConfig'
+import Table from './Table'
+import PacketDetailsViewer from './PacketDetailsViewer'
 
 const Dashboard = ({ packets, config }) => {
   /**
    * selectedPacket will be rendered by PacketDetailsViewer if not empty
    */
-  const [selectedPacket, setSelectedPacket] = useState(null);
+  const [selectedPacket, setSelectedPacket] = useState(null)
 
   /**
    * Toggles selected packet between received packet object or empty object {}
@@ -21,31 +20,33 @@ const Dashboard = ({ packets, config }) => {
    * useCallback hook invokes this functions only when packet is selected in Table.
    */
   const getSelectedPacket = useCallback((packet) => {
-    setSelectedPacket(selectedPacket ? null : packet);
-  }, []);
+    setSelectedPacket(selectedPacket ? null : packet)
+  }, [])
 
   /**
    *  Current config state
    */
   const [currentConfig, setCurrentConfig] = useState(() =>
     config ? mergeConfig(defaultConfig, config) : defaultConfig
-  );
+  )
 
   function mergeConfig(oldConfig, newConfig) {
-    let mergedConfig = {};
+    let mergedConfig = {}
     for (let key in newConfig) {
-      mergedConfig[key] = { ...oldConfig[key], ...newConfig[key] };
+      mergedConfig[key] = { ...oldConfig[key], ...newConfig[key] }
     }
-    return mergedConfig;
+    return mergedConfig
   }
 
   /**
    * Config objects are passed to respective components.
    */
-  const { dashboardConfig, tableConfig, detailsConfig } = currentConfig;
+  const { dashboardConfig, tableConfig, detailsConfig } = currentConfig
+
+  if (!packets) return <h2>No packets provided</h2>
 
   return (
-    <div className="packet-dashboard">
+    <div className='packet-dashboard'>
       {packets && (
         <Table
           getSelectedPacket={getSelectedPacket}
@@ -61,6 +62,6 @@ const Dashboard = ({ packets, config }) => {
         />
       )}
     </div>
-  );
-};
-export default Dashboard;
+  )
+}
+export default Dashboard
