+++ src/lib/components/PacketDetailsViewer.js
