+++ src/lib/constants/defaultConfig.js
