+++ src/lib/index.js
@@ -1,0 +1,3 @@
+import Dashboard from "./components/Dashboard";
+
+export default Dashboard;
